Use OMP Parallel in conductivity module
This commit updates conductivity module to use OMP parallelization
to optimize the performance. It also uses adaptive timestep in calculating
correlation functions to increase the efficiency. After 1000 th frame,
adjacent dt values are 1/100 of the timestamp.
This also includes changes in the conductivity parameter file to calculate
msd value in the direction of msd_vec parameter defined in param.dat file